functions/data-processing-engines/bigquery-executor/requirements.txt
functions-framework==3.*
google-auth
google-cloud-bigquery
grpcio
grpcio-tools
google-api-python-client
google-cloud-dataform